Added offers on Fruit

+++ src/main/java/com/coding/task/ordersystem/model/Fruit.java
@@ -16,7 +16,7 @@ public class Fruit {
         this.fruitName = fruitName;
     }
 
-    public float getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
